+++ src/Moxl/Xec/Payload/Moderated.php
@@ -2,17 +2,17 @@
 
 namespace Moxl\Xec\Payload;
 
-class Retracted extends Payload
+class Moderated extends Payload
 {
     public function handle(?\SimpleXMLElement $stanza = null, ?\SimpleXMLElement $parent = null)
     {
         if ($parent->{'apply-to'} && $parent->{'apply-to'}->attributes()->xmlns == 'urn:xmpp:fasten:0') {
             $message = \App\User::me()->messages()
-                                      ->where('originid', (string)$parent->{'apply-to'}->attributes()->id)
+                                      ->where('stanzaid', (string)$parent->{'apply-to'}->attributes()->id)
                                       ->where('jidfrom', baseJid((string)$parent->attributes()->from))
                                       ->first();
 
-            if ($message && !$message->isMuc()) {
+            if ($message && $message->isMuc()) {
                 $message->retract();
                 $message->save();